docshell/test/mochitest/test_navigate_after_pagehide.html

<!doctype html>
<html>
  <head>
    <title>Test for navigation attempts by scripts in inactive inner window</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  </head>
<body>
<iframe src="dummy_page.html" id="iframe"></iframe>

<script>
"use strict";

add_task(async function() {
  let iframe = document.getElementById("iframe");

  let navigate = iframe.contentWindow.eval(`(function() {
    location.href = "/";
  })`);

  // eslint-disable-next-line @microsoft/sdl/no-insecure-url
  iframe.src = "http://example.com/";
  await new Promise(resolve =>
    iframe.addEventListener("load", resolve, { once: true })
  );

  // This should do nothing. But, importantly, it should especially not crash.
  navigate();

  ok(true, "We didn't crash");
});
</script>
</body>
</html>
